Create journal file
1.	Briefly describe your project. (About 2 sentences.)
Our project is about an online ordering system for a restaurant or a food truck. Where the costumer can order online then come and it will be ready for pick up. In addition, the name of the project is Restaurant Delivery and Pickup System (RDPS).
2.	What is your particular role in the project?
My role in this project is to be the Database guy to link the sign-up information to the SQL database system.
3.	What have you done personally in the two weeks leading up to Feb. 24?
Since I am the new member to the group where my group mates know and have worked on the project, they walked me through it. Moreover, I have been tasked with my part of the project where I have illustrated previously.
4.	Describe any significant progress made by the team as a whole in this time period.
We have assured that our project meets the requirements of this class, and we have redesigned the system. In addition, we have the divided the work between the team members.